refactor folders api
