+++ src/components/Folders/Folders.vue
@@ -47,7 +47,7 @@ export default {
     methods: {
 
         fetchFolders () {
-            Api.fetchFolders()
+            Api.folders.get()
                 .then(response => this.processFolders(response));
         },
