+++ src/utils/api/index.js
@@ -1,11 +1,13 @@
-import Stream from '@/utils/api/stream.js';
 import Account from '@/utils/api/account.js';
 import Conversations from '@/utils/api/conversations.js';
+import Folders from '@/utils/api/folders.js';
 import Messages from '@/utils/api/messages.js';
+import Stream from '@/utils/api/stream.js';
 
 export {
     Account,
     Conversations,
+    Folders,
     Messages,
     Stream
 }
